backend/requirements/base.txt
Django==5.1.6
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-filter==24.3
django-cors-headers==4.6.0
psycopg[binary]==3.2.6
celery==5.5.0
redis==5.2.1
gunicorn==23.0.0
